carbon-api/requirements.txt
fastapi==0.65.2
motor==2.4.0
python-jose==3.3.0